AI小说风格学习生成器/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
chardet==5.2.0
